Organized layout into initial components
