+++ src/components/Layout/AppBar/index.tsx
@@ -1,0 +1,78 @@
+import React from 'react';
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import IconButton from '@material-ui/core/IconButton';
+import Typography from '@material-ui/core/Typography';
+import Badge from '@material-ui/core/Badge';
+import { createStyles, Theme, makeStyles } from '@material-ui/core/styles';
+import MenuIcon from '@material-ui/icons/Menu';
+import AccountCircle from '@material-ui/icons/AccountCircle';
+import MailIcon from '@material-ui/icons/Mail';
+import NotificationsIcon from '@material-ui/icons/Notifications';
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    appBar: {
+        zIndex: theme.zIndex.drawer + 101,
+    },
+    toolBar: {
+        minHeight: 48,
+    },
+    menuButton: {
+        marginLeft: -12,
+        marginRight: 20,
+    },
+    title: {
+        display: 'none',
+        [theme.breakpoints.up('sm')]: {
+            display: 'block',
+        },
+    },
+    grow: {
+        flexGrow: 1,
+    },
+  }),
+);
+
+const AppTopBar: React.FC = () => {
+    const classes = useStyles();
+
+    return (
+        <AppBar position="fixed" className={classes.appBar} >
+          <Toolbar className={classes.toolBar} >
+            
+            {/* Menu Toggle Icon  */}
+            <IconButton className={classes.menuButton} color="inherit" aria-label="Open drawer">
+              <MenuIcon />
+            </IconButton>
+
+            {/* WebSite Name  */}
+            <Typography className={classes.title} variant="h6" color="inherit" noWrap>
+              Material-UI
+            </Typography>
+
+            {/* Separator  */}
+            <div className={classes.grow} />
+
+            <IconButton color="inherit">
+                <Badge badgeContent={4} color="secondary">
+                    <MailIcon />
+                </Badge>
+            </IconButton>
+            
+            <IconButton color="inherit">
+                <Badge badgeContent={17} color="secondary">
+                    <NotificationsIcon />
+                </Badge>
+            </IconButton>
+            
+            <IconButton color="inherit">
+                <AccountCircle />
+            </IconButton>
+
+          </Toolbar>
+        </AppBar>
+    );
+  }
+
+export default AppTopBar;
