+++ src/components/Layout/ContentWrapper/index.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    content: {
+        flexGrow: 1,
+        padding: theme.spacing(3),
+      },
+  })
+);
+
+interface ContentWrapper {
+    children?: React.ReactNode;
+  }
+
+const ContentWrapper: React.FC<ContentWrapper> = props=> {
+  const classes = useStyles();
+
+  return (
+    <div className={classes.content}>
+      <CssBaseline />
+      {props.children}
+    </div>
+  );
+}
+
+export default ContentWrapper;
